tests: Test with a single trigger with a single command
Again, this test comes with its own Yum repository.

However, the files inside are hard links to the files in the repository
for the previous unit tests, so that should save some space hopefully.

I'll try and "reuse" those files (.rpm, .xml, xml.gz) as much as
possible for all unit tests.