Hook into postverifytrans
Turns out posttrans is not the latest hook after the transaction.

Hooking in posttrans, we were doing our stuff before the verification
of the transaction, and in some cases (only one of our unit tests,
but not every time), yum would output a scary warning:
    Warning: RPMDB altered outside of yum.

Hooking after the transaction has been verified is not only a good
idea (after all, if Yum verifies it, there probably is a reason, so
let's wait for everything to have been done properly before doing
our stuff), it also just so happens to get rid of the warning. :)