Sort triggers
This should help running the triggers in a more deterministic order (i.e
two machines getting the same packages as an update will run the same
triggers in the same order).

It's obviously a very naive and imperfect order, but it's nice to have
it deterministic (so all machines fail/succeed in the same way).